Move Pedersen hash functions to their own submodule
